Note that this document should apply to post-4.0 versions of Digital^H^H^H^H^H^H^HTru64 UNIX, but make the note about DU 3.2 a bit speculative (I don't have 3.2 documentation handy to check whether, for example, you can use "doconfig" interactively).
Add a note about using "doconfig" interactively - the Tru64 UNIX FAQ
mentions that in its not on using tcpdump.

Expand on the discussion of "pfconfig" to explain that you're also
enabling "local copy" mode, and to mention the "-a" option to let you
enable local copy and promiscuous mode on all devices, and to explain
*why* you want to enable local copy and promiscuous mode.  Also, in the
example pfconfig command, put the network device name at the end - the
Tru64 UNIX FAQ, and the pfconfig man page, both put it there.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@2937 f5534014-38df-0310-8fa8-9805f1628bb7