Tvbuffify the ISAKMP dissector.
Handle the payloads iteratively rather than recursively, so that the
code that knows how to iterate over payloads is confined to a small
number of places rather than being in every single payload dissector.

Pull the code to dissect the generic payload header into a single
routine, and do it in the code that iterates over payloads rather than
in the dissectors for each payload.

Clean up some other things.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@3085 f5534014-38df-0310-8fa8-9805f1628bb7