Display the printer type/capability bits in the standard way we display bitfields.
Rename "get_space()" to "skip_space()", to indicate what it does, and
have it return FALSE if the first non-blank character it found was a CR
or LF (meaning "end of packet").  If it returns FALSE, stop dissecting
the packet.

In "get_unquoted_string()", search not only for a space, but also for a
tab, CR, or LF; there is no guarantee that there will be any fields in
the packet past the URI (CUPS 1.0[.x] didn't put anything after the URI,
and that's what I'm running on my machine...).


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@3130 f5534014-38df-0310-8fa8-9805f1628bb7