Fix "COOK_ADDR_SELECTOR()" to extract the 2 bits that determine the interpretation of the address fields, in DATA_ADDR_T{1,2,3,4} format.
Clean up a bunch of other macros to enclose their argument in
parentheses.

Clean up "find_header_length()" - it only needs to take the frame
control field of the frame as an argument, and should just check whether
"COOK_ADDR_SELECTOR()" returns DATA_ADDR_T4.

"tofrom_ds" is a value_string table, not a true_false_string table, so
wrap it in "VALS()", not "TFS()".

Don't fetch the header length in N different places in the main
dissector - just fetch it once and use it throughout.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@3582 f5534014-38df-0310-8fa8-9805f1628bb7