Tvbuffify the rlogin dissector.
Get rid of some unneeded includes.

"info_framenum" is set from an Ethereal frame number, and those are
unsigned, so make it unsigned.  No frame will have zero as its frame
number, so start "info_framenum" as 0, not -1.

Fix some long-standing bugs in the packet processing found either as a
result of code inspection in the process of tvbuffifying or doing
regression tests.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@3637 f5534014-38df-0310-8fa8-9805f1628bb7