Use "val_to_str()" to translate numerical values to strings, don't invent something that's almost like "value_string" and almost like "val_to_str()".
Split the command-code field into client and server command code fields,
make them enumerated fields, and put that field into the tree when
dissecting it in the header, not when putting the body into the tree.

Put the body of both unknown client and server requests into the tree,
and just label it as "Body", as is done with the body of known requests.

Display the status code in the same fashion in all places where it's put
into the protocol tree.

When dissecting a SRV_MULTI message, keep dissecting until we run out of
submessages - don't quit when we run out of data in the packet, as that
means we won't throw an exception and won't take the frame as short.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@3646 f5534014-38df-0310-8fa8-9805f1628bb7