Fix the version value for WCCPv2.
Use "proto_tree_add_item()" where possible, and use it, rather than
"proto_tree_add_text()", to put the WCCP version into the packet.  Make
the version field hex, so the V2 version shows up in a reasonable form.

Show the flags in bit-set capabilities as Boolean bitfields.

Check the length field of capabilities for sanity, and use the length
field to step to the next capability rather than hardwiring in 8.

In the top-level tree item for a capability element, show the type of
element.

Use the correct offsets when putting stuff into the subtree for a
capability element.

Dissect the values of unknown capabilities as opaque arrays of bytes,
not as 4-byte integers.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@3816 f5534014-38df-0310-8fa8-9805f1628bb7