"value_string" arrays must end with a terminator entry.
If "get_hex_uint()" supplies a "next_offset" equal to the offset fed
into it, it found no hex digits; don't put the value into the tree if
that's the case.

If "get_unquoted_string()" or "get_quoted_string()" returns NULL, the
string separator/terminator wasn't found; don't put the value into the
tree if that's the case.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@3947 f5534014-38df-0310-8fa8-9805f1628bb7