Base decisions on whether to dump the word parameters on the word count value being non-zero, not on whether the error code is zero.  Don't bother passing the error code to dissectors for particular SMBs, as they don't need to use it.
In "get_unicode_or_ascii_string()", when aligning to an even boundary,
align to an even boundary in the SMB message, not in the packet as a
whole - there's no guarantee that there are an even number of bytes in
the frame before the SMB message.

In the Info column, mark the packet as a request or response based on
the request/response bit in the Flags field, not on the matched port -
for NBIPX, the source and destination ports (IPX sockets) may be the
same, so you may not be able to determine whether it's a request or a
response based on that.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@3965 f5534014-38df-0310-8fa8-9805f1628bb7