For requests and responses for unknown interfaces, put an entry into the DCE RPC protocol tree for the stub data.
Use the counts of context items and transfer syntax items when
dissecting a bind or alter context PDU.

In bind and alter context PDUs, create the conversation, attach the
context ID and interface to it, and put the interface information into
the Info column as soon as the first context item is dissected, so that
if we get an exception after that, we've still processed the context ID
and interface information.

Use the count of results when dissecting a bind ack PDU.

In bind ack PDUs, dissect the transfer syntax and syntax version fields,
and put the opnum and context ID information into the Info column as
soon as it's dissected.

When dissecting a connection-oriented request or response, don't make
the tvbuff the full fragment length if we don't have that much data in
the frame being dissected.  (We should do TCP reassembly there,
eventually.)

In connection-oriented response PDUs, put the opnum and context ID
information into the Info column as soon as it's dissected.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@3982 f5534014-38df-0310-8fa8-9805f1628bb7