Use "val_to_str()", rather than "match_strval()", in "socket_text()".
Make the source and destination socket fields enumerated types, so we
don't have to use "proto_tree_add_uint_format()" on them, and so that
you can match on them by service name.

Use lower-case letters when formatting the hex value of sockets; that's
what's done with fields not added with "proto_tree_add_XXX_format".


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@4046 f5534014-38df-0310-8fa8-9805f1628bb7