Put the stuff to handle file data in read and write requests into a common subroutine.
Label the "total data length" field in Write Raw and Write Multiplexed
requests as such.

Dissect the IPC State/Device State field of Open and X requests and NT
Create and X requests ass per the stuff on page 67 of

	ftp://ftp.microsoft.com/developr/drg/CIFS/dosextp.txt

Make a variable that doesn't need to be static not static.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@4179 f5534014-38df-0310-8fa8-9805f1628bb7