Don't save a full "smb_info" structure in the hash tables for requests and replies, just save a structure holding that information that can't be derived from the contents of one of the SMBs.
Don't save anything at all for NT Cancel requests - they have the same
TID/PID/MID/UID as the SMB being cancelled, and you want the information
for that request used when dissecting the NT Cancel (so it gets the
number of the frame containing the request being cancelled) and when
dissecting the reply to the request being cancelled.

Get rid of an unused routine.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@4213 f5534014-38df-0310-8fa8-9805f1628bb7