zlib 1.0.8 - the version that comes with X11 - does have "gzseek()", even if it doesn't have "gzgets()", so one might think we could use it by using our own replacement for "gzgets()".
One would be wrong to think so, however, as the "gzseek()" it has
doesn't actually work when reading uncompressed files.

zlib 1.0.9 has "gzgets()", and fixes that bug, so we rever to checking
for "gzgets()" rather than "gzseek()", so that we don't accept pre-1.0.9
versions of zlib, and we get rid of our "gzgets()" replacement.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@4702 f5534014-38df-0310-8fa8-9805f1628bb7